// miniproject1-frontend/types/api.ts
export interface components {
    schemas: {
// ...
        PriceHistoryResponse: {
            /** Format: int64 */
            pharmacyId?: number;
            /** Format: int64 */
            drugId?: number;
            points?: components["schemas"]["PricePoint"][];
        };
        PricePoint: {
            /** Format: date */
            purchasedAt?: string;
            /** Format: int32 */
            price?: number;
            flagged?: boolean;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}